# lib/api-zod/src/generated/types/adminRecipe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminIngredient } from "./adminIngredient";
import type { AdminRecipeStatus } from "./adminRecipeStatus";
import type { AdminRecipeStep } from "./adminRecipeStep";

export interface AdminRecipe {
  id: string;
  title: string;
  description?: string;
  countryId: string;
  countryName?: string;
  region?: string;
  category: string;
  difficulty: string;
  prepTime?: string;
  cookTime?: string;
  servings: number;
  culturalNote?: string;
  image?: string;
  status: AdminRecipeStatus;
  featured?: boolean;
  featuredOrder?: number;
  cookCount?: number;
  tips?: string[];
  ingredients?: AdminIngredient[];
  steps?: AdminRecipeStep[];
  createdAt?: string;
}