# lib/api-zod/src/generated/types/adminRecipeStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRecipeStep {
  id: string;
  title?: string;
  instruction: string;
  instructionFirstSteps?: string;
  instructionChefsTable?: string;
  materials?: string[];
  duration?: number;
}