# lib/api-zod/src/generated/types/adminRecipeUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminIngredient } from "./adminIngredient";
import type { AdminRecipeStep } from "./adminRecipeStep";

export interface AdminRecipeUpdate {
  title?: string;
  description?: string;
  countryId?: string;
  region?: string;
  category?: string;
  difficulty?: string;
  prepTime?: string;
  cookTime?: string;
  servings?: number;
  culturalNote?: string;
  image?: string;
  status?: string;
  featured?: boolean;
  featuredOrder?: number;
  tips?: string[];
  ingredients?: AdminIngredient[];
  steps?: AdminRecipeStep[];
}